# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Data Pad Data api specifications",
        "title": "Data Pad Data Api",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "hyurtseven81@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/v1/projects/{project_id}/projectmessages": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projectmessage"
                ],
                "summary": "get all projectmessages under a project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project Id",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "string as JSON / used for filtering the results",
                        "name": "filter",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "string as JSON / used for projecting the results",
                        "name": "select",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "string as JSON / used for sorting the results",
                        "name": "sort",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "used for skipping the results",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "used for limitting the results",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ProjectMessage result",
                        "schema": {
                            "$ref": "#/definitions/utils.DataResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "500": {
                        "description": "error details",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projectmessage"
                ],
                "summary": "insert new projectmessage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project Id",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ProjectMessage Info",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/projectmessage.ProjectMessage"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "OK",
                        "schema": {
                            "type": "integer"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/v1/projects/{project_id}/projectmessages/{project_message_id}": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projectmessage"
                ],
                "summary": "get a projectmessage under a project",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project Id",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ProjectMessage Id",
                        "name": "project_message_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ProjectMessage result",
                        "schema": {
                            "$ref": "#/definitions/projectmessage.ProjectMessage"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "404": {
                        "description": "Document not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projectmessage"
                ],
                "summary": "update a projectmessage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project Id",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ProjectMessage Id",
                        "name": "project_message_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "ProjectMessage Info",
                        "name": "message",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/projectmessage.ProjectMessage"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ProjectMessage result",
                        "schema": {
                            "$ref": "#/definitions/utils.DataResult"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "projectmessage"
                ],
                "summary": "delete a projectmessage",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Project Id",
                        "name": "project_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "ProjectMessage Id",
                        "name": "project_message_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Empty response",
                        "schema": {
                            "type": "int"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "projectmessage.ProjectMessage": {
            "type": "object",
            "required": [
                "from",
                "project_id",
                "text"
            ],
            "properties": {
                "_id": {
                    "type": "string"
                },
                "from": {
                    "type": "string"
                },
                "project_id": {
                    "type": "string"
                },
                "sys": {
                    "$ref": "#/definitions/projectmessage.Sys"
                },
                "text": {
                    "type": "string"
                },
                "to": {
                    "type": "string"
                }
            }
        },
        "projectmessage.Sys": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "created_by": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "updated_by": {
                    "type": "string"
                }
            }
        },
...
        "utils.Data": {
            "type": "object",
            "properties": {
                "count": {
                    "type": "integer"
                },
                "items": {
                    "type": "object"
                }
            }
        },
        "utils.DataResult": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/utils.Data"
                }
            }
        }
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}